Changes left over text to avenir, changes to clickable area to navigate
makes it so full blue area is used to navigate to create booking and find booking along with the text itself, was only image area itself prior. also may or may not have broken some layout as i rearanged some of the boxes on the 'booking home' page that were misaligned before realising it may have been the same story as the logo and is only broken on my screen, but hopefully not.